+++ src/routers/api/index.api.js
@@ -1,7 +1,7 @@
 import CustomRouter from "../CustomRouter.js";
 import { usersRouter } from "./users.api.js";
 import { productsRouter } from "./products.api.js";
-import { cartsRouter } from "./cart.api.js";
+import { cartsRouter } from "./carts.api.js";
 import { sessionsRouter } from "./sessions.api.js";
 import { ticketsRouter } from "./tickets.api.js";
